Expand sys and INFORMATION_SCHEMA metadata and SRP tables.rs
- Implemented sys.parameters, sys.procedures, sys.functions, sys.table_types.
- Enhanced sys.types and sys.columns to include user-defined table types.
- Implemented missing INFORMATION_SCHEMA views: ROUTINE_COLUMNS, TABLE_PRIVILEGES, COLUMN_PRIVILEGES, VIEW_TABLE_USAGE, VIEW_COLUMN_USAGE.
- Refactored crates/tsql_core/src/executor/metadata/sys/tables.rs into a module for better maintainability (SRP).
- Expanded differential compatibility runner queries.
- Fixed inconsistencies in system_type_id across metadata views.
- Added comprehensive integration tests in sys_metadata_repro.rs.
