+++ crates/tsql_core/src/executor/metadata/mod.rs
@@ -2,8 +2,11 @@ mod info_schema_columns;
 mod info_schema_constraints;
 mod info_schema_dispatch;
 mod info_schema_empty;
+mod info_schema_privileges;
+mod info_schema_routine_columns;
 mod info_schema_routines;
 mod info_schema_tables;
+mod info_schema_views;
 mod sys;
 pub(crate) mod system_vars;
