--- crates/tsql_core/src/executor/metadata/sys/tables.rs
@@ -1,859 +1,0 @@
-use super::super::VirtualTable;
-use super::super::{builtin_types_rows, system_type_id, type_max_length, virtual_table_def};
-use crate::catalog::Catalog;
-use crate::storage::StoredRow;
-use crate::types::{DataType, Value};
-
-pub(crate) struct SysSchemas;
-pub(crate) struct SysDatabases;
-pub(crate) struct SysSysDatabases;
-pub(crate) struct SysConfigurations;
-pub(crate) struct SysTables;
-pub(crate) struct SysColumns;
-pub(crate) struct SysAllColumns;
-pub(crate) struct SysDataSpaces;
-pub(crate) struct SysExtendedProperties;
-pub(crate) struct SysIndexColumns;
-pub(crate) struct SysForeignKeyColumns;
-pub(crate) struct SysXmlSchemaCollections;
-pub(crate) struct SysXmlIndexes;
-pub(crate) struct SysTableTypes;
-pub(crate) struct SysEdgeConstraints;
-pub(crate) struct SysAssemblyModules;
-pub(crate) struct SysTriggers;
-pub(crate) struct SysSqlModules;
-pub(crate) struct SysSystemSqlModules;
-pub(crate) struct SysStats;
-pub(crate) struct SysTypes;
-pub(crate) struct SysServerPrincipals;
-
-#[derive(Clone, Copy)]
-struct DatabaseRow {
-    id: i32,
-    name: &'static str,
-    compatibility_level: u8,
-    recovery_model: &'static str,
-}
-
-const CATALOG_DATABASES: &[DatabaseRow] = &[
-    DatabaseRow {
-        id: 1,
-        name: "master",
-        compatibility_level: 160,
-        recovery_model: "FULL",
-    },
-    DatabaseRow {
-        id: 2,
-        name: "tempdb",
-        compatibility_level: 160,
-        recovery_model: "SIMPLE",
-    },
-    DatabaseRow {
-        id: 3,
-        name: "model",
-        compatibility_level: 160,
-        recovery_model: "FULL",
-    },
-    DatabaseRow {
-        id: 4,
-        name: "msdb",
-        compatibility_level: 160,
-        recovery_model: "FULL",
-    },
-];
-
-const USER_DATABASES: &[DatabaseRow] = &[
-    // The playground database is a user database, not a system database.
-    DatabaseRow {
-        id: 5,
-        name: "tsql_wasm",
-        compatibility_level: 160,
-        recovery_model: "FULL",
-    },
-];
-
-fn catalog_databases() -> impl Iterator<Item = &'static DatabaseRow> {
-    CATALOG_DATABASES.iter().chain(USER_DATABASES.iter())
-}
-
-impl VirtualTable for SysSchemas {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "schemas",
-            vec![
-                ("schema_id", DataType::Int, false),
-                ("name", DataType::VarChar { max_len: 128 }, false),
-            ],
-        )
-    }
-
-    fn rows(&self, catalog: &dyn Catalog) -> Vec<StoredRow> {
-        catalog
-            .get_schemas()
-            .iter()
-            .map(|s| StoredRow {
-                values: vec![Value::Int(s.id as i32), Value::VarChar(s.name.clone())],
-                deleted: false,
-            })
-            .collect()
-    }
-}
-
-impl VirtualTable for SysDatabases {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "databases",
-            vec![
-                ("database_id", DataType::Int, false),
-                ("name", DataType::VarChar { max_len: 128 }, false),
-                ("source_database_id", DataType::Int, true),
-                ("owner_sid", DataType::VarBinary { max_len: 85 }, true),
-                ("create_date", DataType::DateTime, false),
-                ("compatibility_level", DataType::TinyInt, false),
-                ("collation_name", DataType::VarChar { max_len: 128 }, false),
-                ("state", DataType::TinyInt, false),
-                ("state_desc", DataType::VarChar { max_len: 60 }, false),
-                ("user_access", DataType::TinyInt, false),
-                ("user_access_desc", DataType::VarChar { max_len: 60 }, false),
-                ("is_read_only", DataType::Bit, false),
-                ("is_fulltext_enabled", DataType::Bit, false),
-                ("recovery_model", DataType::TinyInt, false),
-                (
-                    "recovery_model_desc",
-                    DataType::VarChar { max_len: 60 },
-                    false,
-                ),
-                ("is_auto_close_on", DataType::Bit, false),
-                ("is_auto_shrink_on", DataType::Bit, false),
-                ("is_in_standby", DataType::Bit, false),
-                ("is_distributor", DataType::Bit, false),
-                ("is_cleanly_shutdown", DataType::Bit, false),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        let created = Value::DateTime(
-            chrono::NaiveDate::from_ymd_opt(2026, 1, 1)
-                .unwrap()
-                .and_hms_opt(0, 0, 0)
-                .unwrap(),
-        );
-        catalog_databases()
-            .map(|db| StoredRow {
-                values: vec![
-                    Value::Int(db.id),
-                    Value::VarChar(db.name.to_string()),
-                    Value::Null,
-                    Value::VarBinary(vec![0x01]),
-                    created.clone(),
-                    Value::TinyInt(db.compatibility_level),
-                    Value::VarChar("SQL_Latin1_General_CP1_CI_AS".to_string()),
-                    Value::TinyInt(0),
-                    Value::VarChar("ONLINE".to_string()),
-                    Value::TinyInt(0),
-                    Value::VarChar("MULTI_USER".to_string()),
-                    Value::Bit(false),
-                    Value::Bit(false),
-                    Value::TinyInt(match db.recovery_model {
-                        "SIMPLE" => 3,
-                        "BULK_LOGGED" => 2,
-                        _ => 1,
-                    }),
-                    Value::VarChar(db.recovery_model.to_string()),
-                    Value::Bit(false),
-                    Value::Bit(false),
-                    Value::Bit(false),
-                    Value::Bit(false),
-                    Value::Bit(db.name.eq_ignore_ascii_case("master")),
-                ],
-                deleted: false,
-            })
-            .collect()
-    }
-}
-
-impl VirtualTable for SysConfigurations {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "configurations",
-            vec![
-                ("configuration_id", DataType::Int, false),
-                ("name", DataType::VarChar { max_len: 128 }, false),
-                ("value", DataType::Int, false),
-                ("value_in_use", DataType::Int, false),
-                ("minimum", DataType::Int, false),
-                ("maximum", DataType::Int, false),
-                ("is_dynamic", DataType::Bit, false),
-                ("is_advanced", DataType::Bit, false),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        vec![StoredRow {
-            // configuration_id=16384 used by SSMS probe (contained db auth).
-            values: vec![
-                Value::Int(16384),
-                Value::VarChar("contained database authentication".to_string()),
-                Value::Int(0),
-                Value::Int(0),
-                Value::Int(0),
-                Value::Int(1),
-                Value::Bit(true),
-                Value::Bit(true),
-            ],
-            deleted: false,
-        }]
-    }
-}
-
-impl VirtualTable for SysSysDatabases {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "sysdatabases",
-            vec![
-                ("name", DataType::VarChar { max_len: 128 }, false),
-                ("dbid", DataType::SmallInt, false),
-                ("sid", DataType::VarBinary { max_len: 85 }, true),
-                ("mode", DataType::SmallInt, false),
-                ("status", DataType::Int, false),
-                ("status2", DataType::Int, false),
-                ("crdate", DataType::DateTime, false),
-                ("cmptlevel", DataType::TinyInt, false),
-                ("filename", DataType::VarChar { max_len: 260 }, true),
-                ("version", DataType::Int, false),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        let created = Value::DateTime(
-            chrono::NaiveDate::from_ymd_opt(2026, 1, 1)
-                .unwrap()
-                .and_hms_opt(0, 0, 0)
-                .unwrap(),
-        );
-        catalog_databases()
-            .map(|db| StoredRow {
-                values: vec![
-                    Value::VarChar(db.name.to_string()),
-                    Value::SmallInt(db.id as i16),
-                    Value::VarBinary(vec![0x01]),
-                    Value::SmallInt(0),
-                    Value::Int(0),
-                    Value::Int(0),
-                    created.clone(),
-                    Value::TinyInt(db.compatibility_level),
-                    Value::Null,
-                    Value::Int(0),
-                ],
-                deleted: false,
-            })
-            .collect()
-    }
-}
-
-impl VirtualTable for SysTables {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "tables",
-            vec![
-                ("object_id", DataType::Int, false),
-                ("name", DataType::VarChar { max_len: 128 }, false),
-                ("schema_id", DataType::Int, false),
-                ("principal_id", DataType::Int, true),
-                ("type", DataType::Char { len: 2 }, false),
-                ("type_desc", DataType::VarChar { max_len: 60 }, false),
-                ("create_date", DataType::DateTime, false),
-                ("modify_date", DataType::DateTime, false),
-                ("is_memory_optimized", DataType::Bit, false),
-                ("is_ms_shipped", DataType::Bit, false),
-                ("is_filetable", DataType::Bit, false),
-                ("temporal_type", DataType::TinyInt, false),
-                ("is_external", DataType::Bit, false),
-                ("is_node", DataType::Bit, false),
-                ("is_edge", DataType::Bit, false),
-                ("ledger_type", DataType::Int, true),
-                ("is_dropped_ledger_table", DataType::Bit, false),
-                ("durability", DataType::TinyInt, false),
-                ("durability_desc", DataType::VarChar { max_len: 60 }, false),
-                ("history_table_id", DataType::Int, true),
-            ],
-        )
-    }
-
-    fn rows(&self, catalog: &dyn Catalog) -> Vec<StoredRow> {
-        let created = Value::DateTime(
-            chrono::NaiveDate::from_ymd_opt(2026, 1, 1)
-                .unwrap()
-                .and_hms_opt(0, 0, 0)
-                .unwrap(),
-        );
-        catalog
-            .get_tables()
-            .iter()
-            .map(|t| StoredRow {
-                values: vec![
-                    Value::Int(t.id as i32),
-                    Value::VarChar(t.name.clone()),
-                    Value::Int(t.schema_id as i32),
-                    Value::Null,
-                    Value::Char("U ".to_string()),
-                    Value::VarChar("USER_TABLE".to_string()),
-                    created.clone(),
-                    created.clone(),
-                    Value::Bit(false),
-                    Value::Bit(false),
-                    Value::Bit(false),
-                    Value::TinyInt(0),
-                    Value::Bit(false),
-                    Value::Bit(false),
-                    Value::Bit(false),
-                    Value::Int(0),
-                    Value::Bit(false),
-                    Value::TinyInt(0), // durability (SCHEMA_AND_DATA)
-                    Value::VarChar("SCHEMA_AND_DATA".to_string()), // durability_desc
-                    Value::Null,       // history_table_id
-                ],
-                deleted: false,
-            })
-            .collect()
-    }
-}
-
-impl VirtualTable for SysColumns {
-    fn definition(&self) -> crate::catalog::TableDef {
-        column_table_def("columns", false)
-    }
-
-    fn rows(&self, catalog: &dyn Catalog) -> Vec<StoredRow> {
-        column_rows(catalog, false)
-    }
-}
-
-impl VirtualTable for SysAllColumns {
-    fn definition(&self) -> crate::catalog::TableDef {
-        column_table_def("all_columns", true)
-    }
-
-    fn rows(&self, catalog: &dyn Catalog) -> Vec<StoredRow> {
-        column_rows(catalog, true)
-    }
-}
-
-impl VirtualTable for SysDataSpaces {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "data_spaces",
-            vec![
-                ("data_space_id", DataType::Int, false),
-                ("name", DataType::VarChar { max_len: 128 }, false),
-                ("type", DataType::Char { len: 2 }, false),
-                ("type_desc", DataType::VarChar { max_len: 60 }, false),
-                ("is_default", DataType::Bit, false),
-                ("is_system", DataType::Bit, false),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        vec![StoredRow {
-            values: vec![
-                Value::Int(1),
-                Value::VarChar("PRIMARY".to_string()),
-                Value::Char("FG".to_string()),
-                Value::VarChar("ROWS_FILEGROUP".to_string()),
-                Value::Bit(true),
-                Value::Bit(false),
-            ],
-            deleted: false,
-        }]
-    }
-}
-
-impl VirtualTable for SysExtendedProperties {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "extended_properties",
-            vec![
-                ("major_id", DataType::Int, false),
-                ("minor_id", DataType::Int, false),
-                ("class", DataType::Int, false),
-                ("name", DataType::VarChar { max_len: 128 }, false),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        Vec::new()
-    }
-}
-
-impl VirtualTable for SysIndexColumns {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "index_columns",
-            vec![
-                ("object_id", DataType::Int, false),
-                ("index_id", DataType::Int, false),
-                ("index_column_id", DataType::Int, false),
-                ("column_id", DataType::Int, false),
-                ("is_included_column", DataType::Bit, false),
-            ],
-        )
-    }
-
-    fn rows(&self, catalog: &dyn Catalog) -> Vec<StoredRow> {
-        let mut rows = Vec::new();
-        for idx in catalog.get_indexes() {
-            for (ordinal, col_id) in idx.column_ids.iter().enumerate() {
-                rows.push(StoredRow {
-                    values: vec![
-                        Value::Int(idx.table_id as i32),
-                        Value::Int(idx.id as i32),
-                        Value::Int((ordinal + 1) as i32),
-                        Value::Int(*col_id as i32),
-                        Value::Bit(false),
-                    ],
-                    deleted: false,
-                });
-            }
-        }
-        rows
-    }
-}
-
-impl VirtualTable for SysForeignKeyColumns {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "foreign_key_columns",
-            vec![
-                ("parent_object_id", DataType::Int, false),
-                ("parent_column_id", DataType::Int, false),
-                ("referenced_object_id", DataType::Int, false),
-                ("referenced_column_id", DataType::Int, false),
-            ],
-        )
-    }
-
-    fn rows(&self, catalog: &dyn Catalog) -> Vec<StoredRow> {
-        let mut rows = Vec::new();
-        for table in catalog.get_tables() {
-            for fk in &table.foreign_keys {
-                let ref_schema = fk.referenced_table.schema_or_dbo();
-                let Some(ref_table) = catalog.find_table(ref_schema, &fk.referenced_table.name)
-                else {
-                    continue;
-                };
-                for (i, parent_col_name) in fk.columns.iter().enumerate() {
-                    let Some(parent_col) = table
-                        .columns
-                        .iter()
-                        .find(|c| c.name.eq_ignore_ascii_case(parent_col_name))
-                    else {
-                        continue;
-                    };
-                    let ref_col_name = fk.referenced_columns.get(i).unwrap_or(parent_col_name);
-                    let Some(ref_col) = ref_table
-                        .columns
-                        .iter()
-                        .find(|c| c.name.eq_ignore_ascii_case(ref_col_name))
-                    else {
-                        continue;
-                    };
-                    rows.push(StoredRow {
-                        values: vec![
-                            Value::Int(table.id as i32),
-                            Value::Int(parent_col.id as i32),
-                            Value::Int(ref_table.id as i32),
-                            Value::Int(ref_col.id as i32),
-                        ],
-                        deleted: false,
-                    });
-                }
-            }
-        }
-        rows
-    }
-}
-
-impl VirtualTable for SysXmlSchemaCollections {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "xml_schema_collections",
-            vec![
-                ("xml_collection_id", DataType::Int, false),
-                ("schema_id", DataType::Int, false),
-                ("name", DataType::VarChar { max_len: 128 }, false),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        Vec::new()
-    }
-}
-
-impl VirtualTable for SysXmlIndexes {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "xml_indexes",
-            vec![
-                ("object_id", DataType::Int, false),
-                ("index_id", DataType::Int, false),
-                ("xml_index_type", DataType::TinyInt, false),
-                ("secondary_type", DataType::VarChar { max_len: 1 }, true),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        Vec::new()
-    }
-}
-
-impl VirtualTable for SysTableTypes {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "table_types",
-            vec![
-                ("type_table_object_id", DataType::Int, false),
-                ("is_memory_optimized", DataType::Bit, false),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        Vec::new()
-    }
-}
-
-impl VirtualTable for SysEdgeConstraints {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "edge_constraints",
-            vec![
-                ("name", DataType::VarChar { max_len: 128 }, false),
-                ("parent_object_id", DataType::Int, false),
-                ("create_date", DataType::DateTime, false),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        Vec::new()
-    }
-}
-
-impl VirtualTable for SysAssemblyModules {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "assembly_modules",
-            vec![("object_id", DataType::Int, false)],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        Vec::new()
-    }
-}
-
-impl VirtualTable for SysTriggers {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "triggers",
-            vec![
-                ("name", DataType::NVarChar { max_len: 128 }, false),
-                ("object_id", DataType::Int, false),
-                ("parent_class", DataType::TinyInt, false),
-                ("parent_class_desc", DataType::NVarChar { max_len: 60 }, false),
-                ("parent_id", DataType::Int, false),
-                ("type", DataType::Char { len: 2 }, false),
-                ("type_desc", DataType::NVarChar { max_len: 60 }, false),
-                ("create_date", DataType::DateTime, false),
-                ("modify_date", DataType::DateTime, false),
-                ("is_ms_shipped", DataType::Bit, false),
-                ("is_disabled", DataType::Bit, false),
-                ("is_not_for_replication", DataType::Bit, false),
-                ("is_instead_of_trigger", DataType::Bit, false),
-            ],
-        )
-    }
-
-    fn rows(&self, catalog: &dyn Catalog) -> Vec<StoredRow> {
-        let created = Value::DateTime(
-            chrono::NaiveDate::from_ymd_opt(2026, 1, 1)
-                .unwrap()
-                .and_hms_opt(0, 0, 0)
-                .unwrap(),
-        );
-
-        catalog
-            .get_triggers()
-            .iter()
-            .map(|t| {
-                let parent_id = catalog
-                    .object_id(&t.table_schema, &t.table_name)
-                    .unwrap_or(0);
-                StoredRow {
-                    values: vec![
-                        Value::NVarChar(t.name.clone()),
-                        Value::Int(t.object_id),
-                        Value::TinyInt(1), // OBJECT_OR_COLUMN
-                        Value::NVarChar("OBJECT_OR_COLUMN".to_string()),
-                        Value::Int(parent_id),
-                        Value::Char("TR".to_string()),
-                        Value::NVarChar("SQL_TRIGGER".to_string()),
-                        created.clone(),
-                        created.clone(),
-                        Value::Bit(false), // is_ms_shipped
-                        Value::Bit(false), // is_disabled
-                        Value::Bit(false), // is_not_for_replication
-                        Value::Bit(t.is_instead_of),
-                    ],
-                    deleted: false,
-                }
-            })
-            .collect()
-    }
-}
-
-impl VirtualTable for SysSqlModules {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "sql_modules",
-            vec![
-                ("object_id", DataType::Int, false),
-                ("definition", DataType::VarChar { max_len: 8000 }, true),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        Vec::new()
-    }
-}
-
-impl VirtualTable for SysSystemSqlModules {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "system_sql_modules",
-            vec![
-                ("object_id", DataType::Int, false),
-                ("definition", DataType::VarChar { max_len: 8000 }, true),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        Vec::new()
-    }
-}
-
-impl VirtualTable for SysStats {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "stats",
-            vec![
-                ("object_id", DataType::Int, false),
-                ("name", DataType::VarChar { max_len: 128 }, false),
-                ("auto_created", DataType::Bit, false),
-                ("has_filter", DataType::Bit, false),
-            ],
-        )
-    }
-
-    fn rows(&self, catalog: &dyn Catalog) -> Vec<StoredRow> {
-        catalog
-            .get_indexes()
-            .iter()
-            .map(|idx| StoredRow {
-                values: vec![
-                    Value::Int(idx.table_id as i32),
-                    Value::VarChar(idx.name.clone()),
-                    Value::Bit(false),
-                    Value::Bit(false),
-                ],
-                deleted: false,
-            })
-            .collect()
-    }
-}
-
-fn column_table_def(name: &str, include_sparse: bool) -> crate::catalog::TableDef {
-    let mut cols = vec![
-        ("object_id", DataType::Int, false),
-        ("column_id", DataType::Int, false),
-        ("name", DataType::VarChar { max_len: 128 }, false),
-        ("user_type_id", DataType::Int, false),
-        ("system_type_id", DataType::Int, false),
-        ("max_length", DataType::SmallInt, false),
-        ("precision", DataType::TinyInt, false),
-        ("scale", DataType::TinyInt, false),
-        ("is_nullable", DataType::Bit, false),
-        ("is_computed", DataType::Bit, false),
-        ("is_xml_document", DataType::Bit, false),
-        ("is_column_set", DataType::Bit, false),
-        ("xml_collection_id", DataType::Int, true),
-        ("generated_always_type", DataType::TinyInt, false),
-        ("graph_type", DataType::Int, true),
-        ("default_object_id", DataType::Int, false),
-        ("is_dropped_ledger_column", DataType::Bit, false),
-    ];
-    if include_sparse {
-        cols.push(("is_sparse", DataType::Bit, false));
-    }
-    virtual_table_def(name, cols)
-}
-
-fn column_rows(catalog: &dyn Catalog, include_sparse: bool) -> Vec<StoredRow> {
-    fn precision_scale(dt: &DataType) -> (u8, u8) {
-        match dt {
-            DataType::Bit => (1, 0),
-            DataType::TinyInt => (3, 0),
-            DataType::SmallInt => (5, 0),
-            DataType::Int => (10, 0),
-            DataType::BigInt => (19, 0),
-            DataType::Float => (53, 0),
-            DataType::Decimal { precision, scale } => (*precision, *scale),
-            DataType::Money => (19, 4),
-            DataType::SmallMoney => (10, 4),
-            _ => (0, 0),
-        }
-    }
-
-    let mut rows = Vec::new();
-    for t in catalog.get_tables() {
-        for c in &t.columns {
-            let (precision, scale) = precision_scale(&c.data_type);
-            let mut values = vec![
-                Value::Int(t.id as i32),
-                Value::Int(c.id as i32),
-                Value::VarChar(c.name.clone()),
-                Value::Int(system_type_id(&c.data_type)),
-                Value::Int(system_type_id(&c.data_type)),
-                Value::SmallInt(type_max_length(&c.data_type)),
-                Value::TinyInt(precision),
-                Value::TinyInt(scale),
-                Value::Bit(c.nullable),
-                Value::Bit(c.computed_expr.is_some()),
-                Value::Bit(false),
-                Value::Bit(false),
-                Value::Null,
-                Value::TinyInt(0),
-                Value::Null,
-                Value::Int(if c.default.is_some() {
-                    let table_bucket = (t.id % 100_000) as i32;
-                    3_000_000 + table_bucket * 1_000 + c.id as i32
-                } else {
-                    0
-                }),
-                Value::Bit(false),
-            ];
-            if include_sparse {
-                values.push(Value::Bit(false));
-            }
-            rows.push(StoredRow {
-                values,
-                deleted: false,
-            });
-        }
-    }
-    rows
-}
-
-impl VirtualTable for SysTypes {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "types",
-            vec![
-                ("user_type_id", DataType::Int, false),
-                ("system_type_id", DataType::Int, false),
-                ("name", DataType::VarChar { max_len: 128 }, false),
-                ("max_length", DataType::SmallInt, false),
-                ("precision", DataType::TinyInt, false),
-                ("scale", DataType::TinyInt, false),
-                ("is_user_defined", DataType::Bit, false),
-                ("is_assembly_type", DataType::Bit, false),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        builtin_types_rows()
-            .into_iter()
-            .map(|row| {
-                let id = match row.values.first() {
-                    Some(Value::Int(v)) => *v,
-                    _ => 0,
-                };
-                let name = row
-                    .values
-                    .get(1)
-                    .cloned()
-                    .unwrap_or(Value::VarChar(String::new()));
-                let max_length = row.values.get(2).cloned().unwrap_or(Value::SmallInt(0));
-                let precision = row.values.get(3).cloned().unwrap_or(Value::TinyInt(0));
-                let scale = row.values.get(4).cloned().unwrap_or(Value::TinyInt(0));
-                StoredRow {
-                    values: vec![
-                        Value::Int(id),
-                        Value::Int(id),
-                        name,
-                        max_length,
-                        precision,
-                        scale,
-                        Value::Bit(false),
-                        Value::Bit(false),
-                    ],
-                    deleted: false,
-                }
-            })
-            .collect()
-    }
-}
-
-impl VirtualTable for SysServerPrincipals {
-    fn definition(&self) -> crate::catalog::TableDef {
-        virtual_table_def(
-            "server_principals",
-            vec![
-                ("principal_id", DataType::Int, false),
-                ("name", DataType::VarChar { max_len: 128 }, false),
-                ("type", DataType::Char { len: 1 }, false),
-                ("type_desc", DataType::VarChar { max_len: 60 }, false),
-                ("is_disabled", DataType::Bit, false),
-                ("create_date", DataType::DateTime, false),
-                ("modify_date", DataType::DateTime, false),
-                (
-                    "default_database_name",
-                    DataType::VarChar { max_len: 128 },
-                    true,
-                ),
-            ],
-        )
-    }
-
-    fn rows(&self, _catalog: &dyn Catalog) -> Vec<StoredRow> {
-        let created = Value::DateTime(
-            chrono::NaiveDate::from_ymd_opt(2026, 1, 1)
-                .unwrap()
-                .and_hms_opt(0, 0, 0)
-                .unwrap(),
-        );
-        vec![StoredRow {
-            values: vec![
-                Value::Int(1),
-                Value::VarChar("sa".to_string()),
-                Value::Char("S".to_string()),
-                Value::VarChar("SQL_LOGIN".to_string()),
-                Value::Bit(false),
-                created.clone(),
-                created.clone(),
-                Value::VarChar("master".to_string()),
-            ],
-            deleted: false,
-        }]
-    }
-}
