+++ crates/tsql_core/src/executor/metadata/sys/tables/mod.rs
@@ -1,0 +1,46 @@
+mod columns;
+mod databases;
+mod objects_misc;
+mod tables;
+mod types;
+
+pub(crate) use columns::{SysAllColumns, SysColumns};
+pub(crate) use databases::{SysConfigurations, SysDatabases, SysSysDatabases};
+pub(crate) use objects_misc::{
+    SysAssemblyModules, SysDataSpaces, SysEdgeConstraints, SysExtendedProperties,
+    SysForeignKeyColumns, SysIndexColumns, SysServerPrincipals, SysStats, SysTriggers,
+    SysXmlIndexes, SysXmlSchemaCollections, SysSqlModules, SysSystemSqlModules
+};
+pub(crate) use tables::SysTables;
+pub(crate) use types::{SysTableTypes, SysTypes};
+
+use super::super::VirtualTable;
+use super::super::{virtual_table_def};
+use crate::catalog::Catalog;
+use crate::storage::StoredRow;
+use crate::types::{DataType, Value};
+
+pub(crate) struct SysSchemas;
+
+impl VirtualTable for SysSchemas {
+    fn definition(&self) -> crate::catalog::TableDef {
+        virtual_table_def(
+            "schemas",
+            vec![
+                ("schema_id", DataType::Int, false),
+                ("name", DataType::VarChar { max_len: 128 }, false),
+            ],
+        )
+    }
+
+    fn rows(&self, catalog: &dyn Catalog) -> Vec<StoredRow> {
+        catalog
+            .get_schemas()
+            .iter()
+            .map(|s| StoredRow {
+                values: vec![Value::Int(s.id as i32), Value::VarChar(s.name.clone())],
+                deleted: false,
+            })
+            .collect()
+    }
+}
